Update instances of field.required to field.is_required in the LangChain library, particularly in the _create_subset_model function and similar contexts.
Update anyio version and dependencies

The anyio package version is updated from 3.7.1 to 4.4.0. Along with this change, the Python versions requirement is increased to >=3.8, with updates to the file hash. Adjustments have also been made to package dependencies and extras, including requirement specification for exceptiongroup and typing-extensions, as well as updates to Sphinx, coverage and trio version in extras.